requirements.txt
asgiref==3.11.0
Django==6.0
sqlparse==0.5.5
absl-py==2.3.1
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
daphne==4.2.1
django-cors-headers==4.9.0
django-phonenumber-field==8.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
face-recognition==1.3.0
face_recognition_models==0.3.0
flatbuffers==25.9.23
fonttools==4.60.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
jax==0.7.1
jaxlib==0.7.1
kiwisolver==1.4.9
matplotlib==3.10.7
mediapipe==0.10.21
ml_dtypes==0.5.4
msgpack==1.1.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
phonenumbers==9.0.16
pillow==11.3.0
protobuf==4.25.8
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
redis==7.0.0
requests==2.32.5
scipy==1.16.3
sentencepiece==0.2.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.3
stripe==14.0.1
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
zope.interface==8.0.1

